Ready for MIgration Test

+++ Session_21/CarService.Models/Entities/MonthlyLedger.cs
@@ -9,7 +9,7 @@ namespace CarService.Models.Entities;
 public class MonthlyLedger
 {
     public short Year { get; set; }
-    public short Month { get; set; }
+    public sbyte Month { get; set; }
     public decimal Income { get; set; }
     public decimal Expenses { get; set; }
     public decimal Total { get; set; }
